@@ -78,8 +78,7 @@
         <auto-value.version>0.2.0</auto-value.version>
 
         <slf4j.version>1.7.36</slf4j.version>
-        <spring.version>5.3.24</spring.version>
-        <spring-boot.version>2.7.7</spring-boot.version>
+        <spring-boot.version>3.3.1</spring-boot.version>
         <commons-io.version>2.11.0</commons-io.version>
         <commons-exec.version>1.3</commons-exec.version>
         <jackson.version>2.14.1</jackson.version>
